[SPARK-44051][TESTS][PS][CONNECT] Split `pyspark.pandas.tests.connect.data_type_ops.test_parity_num_ops`
### What changes were proposed in this pull request?
Split `pyspark.pandas.tests.connect.data_type_ops.test_parity_num_ops`

### Why are the changes needed?

it is slow

```
Starting test(python3.9): pyspark.pandas.tests.connect.data_type_ops.test_parity_num_ops (temp output: /__w/spark/spark/python/target/fd59d461-ba78-4672-a164-5d0790c57fb0/python3.9__pyspark.pandas.tests.connect.data_type_ops.test_parity_num_ops__gsu_twd5.log)
Finished test(python3.9): pyspark.pandas.tests.connect.data_type_ops.test_parity_num_ops (333s) ... 1 tests were skipped
```

### Does this PR introduce _any_ user-facing change?
no, test-only

### How was this patch tested?
updated CI

Closes #41591 from zhengruifeng/ps_test_split_num_ops.

Authored-by: Ruifeng Zheng <ruifengz@apache.org>
