+++ python/pyspark/pandas/tests/data_type_ops/test_num_ops.py
@@ -50,215 +50,6 @@ class NumOpsTestsMixin:
     def float_psser(self):
         return ps.from_pandas(self.float_pser)
 
-    def test_add(self):
-        pdf, psdf = self.pdf, self.psdf
-        for col in self.numeric_df_cols:
-            pser, psser = pdf[col], psdf[col]
-            self.assert_eq(pser + pser, psser + psser)
-            self.assert_eq(pser + 1, psser + 1)
-            # self.assert_eq(pser + 0.1, psser + 0.1)
-            self.assert_eq(pser + pser.astype(bool), psser + psser.astype(bool))
-            self.assert_eq(pser + True, psser + True)
-            self.assert_eq(pser + False, psser + False)
-
-            for n_col in self.non_numeric_df_cols:
-                if n_col == "bool":
-                    self.assert_eq(pser + pdf[n_col], psser + psdf[n_col])
-                else:
-                    self.assertRaises(TypeError, lambda: psser + psdf[n_col])
-
-    def test_sub(self):
-        pdf, psdf = self.pdf, self.psdf
-        for col in self.numeric_df_cols:
-            pser, psser = pdf[col], psdf[col]
-            self.assert_eq(pser - pser, psser - psser)
-            self.assert_eq(pser - 1, psser - 1)
-            # self.assert_eq(pser - 0.1, psser - 0.1)
-            self.assert_eq(pser - pser.astype(bool), psser - psser.astype(bool))
-            self.assert_eq(pser - True, psser - True)
-            self.assert_eq(pser - False, psser - False)
-
-            for n_col in self.non_numeric_df_cols:
-                if n_col == "bool":
-                    self.assert_eq(pser - pdf[n_col], psser - psdf[n_col])
-                else:
-                    self.assertRaises(TypeError, lambda: psser - psdf[n_col])
-
-    def test_mul(self):
-        pdf, psdf = self.pdf, self.psdf
-        for col in self.numeric_df_cols:
-            pser, psser = pdf[col], psdf[col]
-            self.assert_eq(pser * pser, psser * psser)
-            self.assert_eq(pser * pser.astype(bool), psser * psser.astype(bool))
-            self.assert_eq(pser * True, psser * True)
-            self.assert_eq(pser * False, psser * False)
-
-            if psser.dtype in [int, np.int32]:
-                self.assert_eq(pser * pdf["string"], psser * psdf["string"])
-            else:
-                self.assertRaises(TypeError, lambda: psser * psdf["string"])
-
-            self.assert_eq(pser * pdf["bool"], psser * psdf["bool"])
-
-            self.assertRaises(TypeError, lambda: psser * psdf["datetime"])
-            self.assertRaises(TypeError, lambda: psser * psdf["date"])
-            self.assertRaises(TypeError, lambda: psser * psdf["categorical"])
-
-    def test_truediv(self):
-        pdf, psdf = self.pdf, self.psdf
-        for col in self.numeric_df_cols:
-            pser, psser = pdf[col], psdf[col]
-            if psser.dtype in [float, int, np.int32]:
-                self.assert_eq(pser / pser, psser / psser)
-                self.assert_eq(pser / pser.astype(bool), psser / psser.astype(bool))
-                self.assert_eq(pser / True, psser / True)
-                self.assert_eq(pser / False, psser / False)
-
-            for n_col in self.non_numeric_df_cols:
-                if n_col == "bool":
-                    self.assert_eq(pdf["float"] / pdf[n_col], psdf["float"] / psdf[n_col])
-                else:
-                    self.assertRaises(TypeError, lambda: psser / psdf[n_col])
-
-    def test_floordiv(self):
-        pdf, psdf = self.pdf, self.psdf
-        pser, psser = pdf["float"], psdf["float"]
-        self.assert_eq(pser // pser, psser // psser)
-        self.assert_eq(pser // pser.astype(bool), psser // psser.astype(bool))
-        self.assert_eq(pser // True, psser // True)
-        self.assert_eq(pser // False, psser // False)
-
-        for n_col in self.non_numeric_df_cols:
-            if n_col == "bool":
-                self.assert_eq(pdf["float"] // pdf["bool"], psdf["float"] // psdf["bool"])
-            else:
-                for col in self.numeric_df_cols:
-                    psser = psdf[col]
-                    self.assertRaises(TypeError, lambda: psser // psdf[n_col])
-
-    def test_mod(self):
-        pdf, psdf = self.pdf, self.psdf
-        for col in self.numeric_df_cols:
-            pser, psser = pdf[col], psdf[col]
-            self.assert_eq(pser % pser, psser % psser)
-            self.assert_eq(pser % pser.astype(bool), psser % psser.astype(bool))
-            self.assert_eq(pser % True, psser % True)
-            if col in ["int", "int32"]:
-                self.assert_eq(
-                    pd.Series([np.nan, np.nan, np.nan], dtype=float, name=col), psser % False
-                )
-            else:
-                self.assert_eq(
-                    pd.Series([np.nan, np.nan, np.nan], dtype=pser.dtype, name=col), psser % False
-                )
-
-            for n_col in self.non_numeric_df_cols:
-                if n_col == "bool":
-                    self.assert_eq(pdf["float"] % pdf[n_col], psdf["float"] % psdf[n_col])
-                else:
-                    self.assertRaises(TypeError, lambda: psser % psdf[n_col])
-
-    def test_pow(self):
-        pdf, psdf = self.pdf, self.psdf
-        for col in self.numeric_df_cols:
-            pser, psser = pdf[col], psdf[col]
-            if col in ["float", "float_w_nan"]:
-                self.assert_eq(pser**pser, psser**psser)
-                self.assert_eq(pser ** pser.astype(bool), psser ** psser.astype(bool))
-                self.assert_eq(pser**True, psser**True)
-                self.assert_eq(pser**False, psser**False)
-                self.assert_eq(pser**1, psser**1)
-                self.assert_eq(pser**0, psser**0)
-
-            for n_col in self.non_numeric_df_cols:
-                if n_col == "bool":
-                    self.assert_eq(pdf["float"] ** pdf[n_col], psdf["float"] ** psdf[n_col])
-                else:
-                    self.assertRaises(TypeError, lambda: psser ** psdf[n_col])
-
-    def test_radd(self):
-        pdf, psdf = self.pdf, self.psdf
-        for col in self.numeric_df_cols:
-            pser, psser = pdf[col], psdf[col]
-            self.assert_eq(1 + pser, 1 + psser)
-            # self.assert_eq(0.1 + pser, 0.1 + psser)
-            self.assertRaises(TypeError, lambda: "x" + psser)
-            self.assert_eq(True + pser, True + psser)
-            self.assert_eq(False + pser, False + psser)
-            self.assertRaises(TypeError, lambda: datetime.date(1994, 1, 1) + psser)
-            self.assertRaises(TypeError, lambda: datetime.datetime(1994, 1, 1) + psser)
-
-    def test_rsub(self):
-        pdf, psdf = self.pdf, self.psdf
-        for col in self.numeric_df_cols:
-            pser, psser = pdf[col], psdf[col]
-            self.assert_eq(1 - pser, 1 - psser)
-            # self.assert_eq(0.1 - pser, 0.1 - psser)
-            self.assertRaises(TypeError, lambda: "x" - psser)
-            self.assert_eq(True - pser, True - psser)
-            self.assert_eq(False - pser, False - psser)
-            self.assertRaises(TypeError, lambda: datetime.date(1994, 1, 1) - psser)
-            self.assertRaises(TypeError, lambda: datetime.datetime(1994, 1, 1) - psser)
-
-    def test_rmul(self):
-        pdf, psdf = self.pdf, self.psdf
-        for col in self.numeric_df_cols:
-            pser, psser = pdf[col], psdf[col]
-            self.assert_eq(1 * pser, 1 * psser)
-            # self.assert_eq(0.1 * pser, 0.1 * psser)
-            self.assertRaises(TypeError, lambda: "x" * psser)
-            self.assert_eq(True * pser, True * psser)
-            self.assert_eq(False * pser, False * psser)
-            self.assertRaises(TypeError, lambda: datetime.date(1994, 1, 1) * psser)
-            self.assertRaises(TypeError, lambda: datetime.datetime(1994, 1, 1) * psser)
-
-    def test_rtruediv(self):
-        pdf, psdf = self.pdf, self.psdf
-        for col in self.numeric_df_cols:
-            pser, psser = pdf[col], psdf[col]
-            # self.assert_eq(5 / pser, 5 / psser)
-            # self.assert_eq(0.1 / pser, 0.1 / psser)
-            self.assertRaises(TypeError, lambda: "x" / psser)
-            self.assert_eq((True / pser).astype(float), True / psser, check_exact=False)
-            self.assert_eq((False / pser).astype(float), False / psser)
-            self.assertRaises(TypeError, lambda: datetime.date(1994, 1, 1) / psser)
-            self.assertRaises(TypeError, lambda: datetime.datetime(1994, 1, 1) / psser)
-
-    def test_rfloordiv(self):
-        pdf, psdf = self.pdf, self.psdf
-        for col in self.numeric_df_cols:
-            pser, psser = pdf[col], psdf[col]
-            # self.assert_eq(5 // pser, 5 // psser)
-            # self.assert_eq(0.1 // pser, 0.1 // psser)
-            self.assertRaises(TypeError, lambda: "x" // psser)
-            self.assert_eq((True // pser).astype(float), True // psser)
-            self.assert_eq((False // pser).astype(float), False // psser)
-            self.assertRaises(TypeError, lambda: datetime.date(1994, 1, 1) // psser)
-            self.assertRaises(TypeError, lambda: datetime.datetime(1994, 1, 1) // psser)
-
-    def test_rpow(self):
-        pdf, psdf = self.pdf, self.psdf
-        for col in self.numeric_df_cols:
-            pser, psser = pdf[col], psdf[col]
-            # self.assert_eq(1 ** pser, 1 ** psser)
-            # self.assert_eq(0.1 ** pser, 0.1 ** psser)
-            self.assertRaises(TypeError, lambda: "x" ** psser)
-            self.assert_eq((True**pser).astype(float), True**psser)
-            self.assert_eq((False**pser).astype(float), False**psser)
-            self.assertRaises(TypeError, lambda: datetime.date(1994, 1, 1) ** psser)
-            self.assertRaises(TypeError, lambda: datetime.datetime(1994, 1, 1) ** psser)
-
-    def test_rmod(self):
-        pdf, psdf = self.pdf, self.psdf
-        for col in self.numeric_df_cols:
-            pser, psser = pdf[col], psdf[col]
-            self.assert_eq(1 % pser, 1 % psser)
-            # self.assert_eq(0.1 % pser, 0.1 % psser)
-            self.assert_eq(True % pser, True % psser)
-            self.assert_eq(False % pser, False % psser)
-            self.assertRaises(TypeError, lambda: datetime.date(1994, 1, 1) % psser)
-            self.assertRaises(TypeError, lambda: datetime.datetime(1994, 1, 1) % psser)
-
     def test_and(self):
         psdf = self.psdf
         for col in self.numeric_df_cols:
